+++ app/layout.tsx
@@ -1,14 +1,8 @@
-"use client";
-export const experimental_ppr = true;
 import { ReactNode } from "react";
-import { ClientWalletProvider } from "@/components/ClientWalletProvider";
-import Navbar from "@/components/Navbar";
-import { Footer } from "@/components/ui/Footer";
-import { DataProvider } from "@/contexts/DataContext";
+import { ClientLayout } from "./client-layout";
 import "./globals.css";
 import { Inter } from "next/font/google";
 import { cn } from "@/lib/utils";
-import { SessionProvider } from "next-auth/react";
 import { GoogleAnalytics } from '@next/third-parties/google'
 
 const inter = Inter({
@@ -17,21 +11,20 @@ const inter = Inter({
   variable: "--font-inter",
 });
 
+import { generateMetadata } from "@/lib/metadata";
+
+export const metadata = generateMetadata({
+  title: "Dashboard",
+  description: "Custom page description"
+});
+
 export default function RootLayout({ children }: { children: ReactNode }) {
   const initialTheme = "dark";
 
   return (
     <html lang="en" className={cn(inter.variable, initialTheme)}>
       <body className="min-h-screen bg-background antialiased" cz-shortcut-listen="true">
-        <SessionProvider>
-          <ClientWalletProvider>
-            <DataProvider>
-              <Navbar />
-              <main className="flex-1">{children}</main>
-              <Footer />
-            </DataProvider>
-          </ClientWalletProvider>
-        </SessionProvider>
+        <ClientLayout>{children}</ClientLayout>
       </body>
       <GoogleAnalytics gaId="G-WM6MY8V617" />
     </html>
